Move BytesResource and StreamResource into dedicated files
